Take out the methods of HGCalGeometryMode from HGCalParametersFromDD.cc to enable clean separation of full HGCal and TB HGCal
