+++ Geometry/HGCalCommonData/src/HGCalParametersFromDD.cc
@@ -11,49 +11,6 @@
 //#define EDM_ML_DEBUG
 using namespace geant_units::operators;
 
-namespace {
-  HGCalGeometryMode::GeometryMode getGeometryMode(const char* s, const DDsvalues_type& sv) {
-    DDValue val(s);
-    if (DDfetch(&sv, val)) {
-      const std::vector<std::string>& fvec = val.strings();
-      if (fvec.empty()) {
-        throw cms::Exception("HGCalGeom") << "getGeometryMode::Failed to get " << s << " tag.";
-      }
-
-      HGCalStringToEnumParser<HGCalGeometryMode::GeometryMode> eparser;
-      HGCalGeometryMode::GeometryMode result = (HGCalGeometryMode::GeometryMode)eparser.parseString(fvec[0]);
-      return result;
-    } else {
-      throw cms::Exception("HGCalGeom") << "getGeometryMode::Failed to fetch " << s << " tag";
-    }
-  }
-  HGCalGeometryMode::GeometryMode getGeometryMode(const std::string& s) {
-    HGCalStringToEnumParser<HGCalGeometryMode::GeometryMode> eparser;
-    HGCalGeometryMode::GeometryMode result = (HGCalGeometryMode::GeometryMode)eparser.parseString(s);
-    return result;
-  }
-  HGCalGeometryMode::WaferMode getGeometryWaferMode(const char* s, const DDsvalues_type& sv) {
-    DDValue val(s);
-    if (DDfetch(&sv, val)) {
-      const std::vector<std::string>& fvec = val.strings();
-      if (fvec.empty()) {
-        throw cms::Exception("HGCalGeom") << "getGeometryWaferMode::Failed to get " << s << " tag.";
-      }
-
-      HGCalStringToEnumParser<HGCalGeometryMode::WaferMode> eparser;
-      HGCalGeometryMode::WaferMode result = (HGCalGeometryMode::WaferMode)eparser.parseString(fvec[0]);
-      return result;
-    } else {
-      throw cms::Exception("HGCalGeom") << "getGeometryWaferMode::Failed to fetch " << s << " tag";
-    }
-  }
-  HGCalGeometryMode::WaferMode getGeometryWaferMode(std::string& s) {
-    HGCalStringToEnumParser<HGCalGeometryMode::WaferMode> eparser;
-    HGCalGeometryMode::WaferMode result = (HGCalGeometryMode::WaferMode)eparser.parseString(s);
-    return result;
-  }
-}  // namespace
-
 bool HGCalParametersFromDD::build(const DDCompactView* cpv,
                                   HGCalParameters& php,
                                   const std::string& name,
@@ -78,7 +35,7 @@ bool HGCalParametersFromDD::build(const DDCompactView* cpv,
 #endif
   if (ok) {
     DDsvalues_type sv(fv.mergedSpecifics());
-    php.mode_ = getGeometryMode("GeometryMode", sv);
+    php.mode_ = HGCalGeometryMode::getGeometryMode("GeometryMode", sv);
 #ifdef EDM_ML_DEBUG
     edm::LogVerbatim("HGCalGeom") << "Volume " << name << " GeometryMode " << php.mode_ << ":"
                                   << HGCalGeometryMode::Hexagon << ":" << HGCalGeometryMode::HexagonFull << ":"
@@ -105,7 +62,7 @@ bool HGCalParametersFromDD::build(const DDCompactView* cpv,
       bool ok2 = fv2.firstChild();
       if (ok2) {
         DDsvalues_type sv2(fv2.mergedSpecifics());
-        mode = getGeometryWaferMode("WaferMode", sv2);
+        mode = HGCalGeometryMode::getGeometryWaferMode("WaferMode", sv2);
 #ifdef EDM_ML_DEBUG
         edm::LogVerbatim("HGCalGeom") << "WaferMode " << mode << ":" << HGCalGeometryMode::Polyhedra << ":"
                                       << HGCalGeometryMode::ExtrudedPolygon;
@@ -150,7 +107,7 @@ bool HGCalParametersFromDD::build(const DDCompactView* cpv,
       bool ok2 = fv2.firstChild();
       if (ok2) {
         DDsvalues_type sv2(fv2.mergedSpecifics());
-        mode = getGeometryWaferMode("WaferMode", sv2);
+        mode = HGCalGeometryMode::getGeometryWaferMode("WaferMode", sv2);
         php.nCellsFine_ = static_cast<int>(getDDDValue("NumberOfCellsFine", sv2));
         php.nCellsCoarse_ = static_cast<int>(getDDDValue("NumberOfCellsCoarse", sv2));
         php.waferSize_ = HGCalParameters::k_ScaleFromDDD * getDDDValue("WaferSize", sv2);
@@ -297,7 +254,7 @@ bool HGCalParametersFromDD::build(const cms::DDCompactView* cpv,
 #endif
 
   if (ok) {
-    php.mode_ = getGeometryMode(sv);
+    php.mode_ = HGCalGeometryMode::getGeometryMode(sv);
 #ifdef EDM_ML_DEBUG
     edm::LogVerbatim("HGCalGeom") << "Volume " << name << " GeometryMode " << php.mode_ << ":"
                                   << HGCalGeometryMode::Hexagon << ":" << HGCalGeometryMode::HexagonFull << ":"
@@ -317,7 +274,7 @@ bool HGCalParametersFromDD::build(const cms::DDCompactView* cpv,
     if ((php.mode_ == HGCalGeometryMode::Hexagon) || (php.mode_ == HGCalGeometryMode::HexagonFull)) {
       tempS = fv.get<std::vector<std::string> >(namet, "WaferMode");
       std::string sv2 = (!tempS.empty()) ? tempS[0] : "HGCalGeometryMode::Polyhedra";
-      mode = getGeometryWaferMode(sv2);
+      mode = HGCalGeometryMode::getGeometryWaferMode(sv2);
 #ifdef EDM_ML_DEBUG
       edm::LogVerbatim("HGCalGeom") << "WaferMode " << mode << ":" << HGCalGeometryMode::Polyhedra << ":"
                                     << HGCalGeometryMode::ExtrudedPolygon;
@@ -365,7 +322,7 @@ bool HGCalParametersFromDD::build(const cms::DDCompactView* cpv,
 
       tempS = fv.get<std::vector<std::string> >(namet, "WaferMode");
       std::string sv2 = (!tempS.empty()) ? tempS[0] : "HGCalGeometryMode::ExtrudedPolygon";
-      mode = getGeometryWaferMode(sv2);
+      mode = HGCalGeometryMode::getGeometryWaferMode(sv2);
       tempD = fv.get<std::vector<double> >(namet, "NumberOfCellsFine");
       php.nCellsFine_ = static_cast<int>(tempD[0]);
       tempD = fv.get<std::vector<double> >(namet, "NumberOfCellsCoarse");
